주문API - 주문테이블, 주문아이템 테이블 분리완료

+++ src/main/java/com/project/starcoffee/controller/request/pay/BalanceRequest.java
@@ -16,5 +16,5 @@ public class BalanceRequest {
     @NotNull
     private UUID cardId;
     @NotNull
-    private int finalPrice;
+    private long finalPrice;
 }
